Initial commit: Basic functionality working
This is a demo to see how to accomplish the basic drawing functionality
in a browser, as a single-page application. Already working are multiple
drawing tools, erase and remove tools, and unlimited undo/redo (with
deltas to limit memory usage). Currently the toolbar is completely
hard-coded, but I have ideas about how to make it more customizable.