New work on a modular/customizable toolbar design
Changes:
- Add an Ellipse tool, rename some other tools.
- Refactor all drawing tools, and add a tool-button class corresponding
to each one. (Roughly MVC, with the button class as the view and the
tool class as the controller.)

Working now:
- Basic buttons for each tool, and for selecting color, opacity, stroke
width, and dash pattern. No other more dynamic widgets for these yet,
but that may happen in the future.
- Tools can be specified either in an HTML document, using custom
elements (Web Components), or by adding these components via Javascript.
- The buttons all work, basically. That is, you can use them to select
various tools, and set the properties of those tools, and it all follows
the rules laid out in the ToolbarDesign document.

Not working yet:
- So far no JSON configuration, but that should be fairly easy to add
later with this design.
- Button appearance is not updated to reflect any choices.
- Some buttons should be disabled based on tool selection.
- Haven't even tested the pass-through mode! That needs to be done.
- No keyboard navigation/shortcuts yet.