Get all basic toolbar functionality working
Changes:
- Refactor all drawing tools (again), so they now encapsulate a
paper.Tool, instead of subclassing it. This allows us to use Javascript
classes, rather than constructor functions/prototypes, so it's more
consistent with the rest of the project. It also allows these tool
objects to be constructed early, before paperJS is initialized.
- Refactor ButtonGroup to WidgetGroup, with two subclasses for
PropertyWidgets and ToolButtons.
- Made properties controlled by a properties object, which is
essentially a dictionary of property names to PropertyWidgetGroups.
- Although button appearances are still hard-coded, there's now nothing
in them that needs to be, despite making their appearance variable.
(Thanks to the magic of CSS custom properties!)

Working now:
- PropertyButtons are now disabled as appropriate, based on the
currently selected tool.
- ToolButtons now update their appearance as appropriate, based on the
currently selected properties.
- Tools can still be specified in an HTML document using custom
elements, but this will likely change soon. Longer term, this will
probably always be configured via a JSON object.

Not working yet:
- JSON configuration
- Keyboard navigation/shortcuts, and the ability to control the whole
app entirely via keyboard shortcuts if the toolbar is hidden. Hopefully
this will not require a major refactor.
- Pass-through mode!