Simplify button code and eliminate many classes
Major refactor to clean up the code, get rid of all the custom elements
(Web Components) for the buttons, and also make the whole system more
configurable.

Changes:
- Got rid of all custom elements for the buttons, and replaced them with
one big factory function. I might change this, however, because it broke
the ability to use a toolbar configured in HTML. There's an easy fix to
bring that back without custom elements, but another way would be to
implement just one custom element that covers all the buttons.
- Got rid of the ToolButtonGroup class, because it's intended to be a
singleton, and it barely differed from the WidgetGroup class anyway.
- All setup code (i.e. *everything* other than function definitions and
class definitions, and the *declaration* (not setting) of exactly five
module-level state variables) now happens in a function, called
initialize().

Working now:
- UI works (or should work) exactly as in previous commit.
- JSON configuration of toolbar is now possible! So far, we don't
actually load it as a JSON file, but it's all stored in a Javascript
object. HTML configuration is slightly broken, but that should be
temporary.
- The appearance of all the widgets is also now stored in a big object,
so the whole UI is theme-able. Right now this theme is split between
this (essentially JSON) object, and the CSS file*, but eventually the
aim is to move all the relevant CSS into the same JSON object.

Not working yet:
- Keyboard navigation/shortcuts.
- Pass-through mode!!