Fix HTML configuration, implement passthrough mode
Changes:
- Got HTML configuration working again, and did so via custom elements,
but in a much simpler way than before. Now there's just one custom
button, and rather than using a bunch of subclasses, we just specify the
type of button via the "data-type" attribute.
- Fixed minor bug to allow de-selecting a property when there's no
widget with that value.
- Implemented the pass-through tool finally! Incredibly simple.
- Created some new config files and a theme file. We're still not
actually using any of these files, yet. Also, they're in Hjson format,
which is a more human-readable superset of JSON, that can be compiled to
proper JSON with a simple utility.

Working now:
- Pass-through tool! This has been tested with static content
underneath, as well as an interactive Plotly graph.

Not working yet:
- Keyboard navigation/shortcuts.
- Ability to configure toolbar at top/bottom/left/right.
- Internal styles are still handled externally. This will eventually be
incorporated into the theme object.