Make the whole Notetaker into a web component!
Changes:
- Wrapped the whole app up into a single custom element, called
note-taker. Now putting this into any web page should be as simple as
including the appropriate scripts and sticking
<note-taker></note-taker> into the HTML somewhere.
- Fixed a bug that affected undo/redo, caused by a problem with how the
EraseTool was handling closed paths (RectangleTool and EllipseTool).
- Modified EllipseTool to only create circles (nicely) if snapTolerance
is set to a negative value.
- Cleaned up code in several places, especially in the constructor and
activate/deactivate methods of the base NotetakerTool class. This also
resolved a bug that never appeared, but could have in rare cases. NOTE:
This new code *should* be more solid than the previous version,
especially in terms of handling edge cases, but it has not been
extensively tested yet.

Working now:
- <note-taker> web component (custom element)
- All styling of the note-taker itself (internal styling) is done
internally, and is encapsulated in the shadow DOM.

Not working yet:
- Passing configuration and theming parameters as attributes on the
<note-taker> element.
- Loading configuration JSON and theme JSON from external files.
- Ability to configure toolbar at top/bottom/left/right.
- Keyboard navigation/shortcuts.