Fetch configuration and theme from external JSON
Changes:
- Updated the note-taker element setup code to take two attributes
(config-href, and optional theme-href) that will fetch JSON data for the
configuration and the theme. The configuration can include a theme URL
within it, so only the configuration attribute needs to be specified.
Also, the result is cached in a module-global variable, so in case
multiple note-taker elements are used, only the first one needs to
specify the configuration and/or theme. Implementation uses fetch, via
async/await.
- Updated the drawing tools to make all parameters (snap for LineTool,
RectangleTool, and EllipseTool, and simplify for PenTool and EraseTool)
properties, so they can be specified in the config file, and potentially
changed via ToolWidgets.
- Added an extra condition to LineTool and RectangleTool (like was
recently done for EllipseTool) so that when snap is negative, the tool
always snaps. That is, RectangleTool always draws squares, and LineTool
always draws lines at multiples of pi/4.
- Cleaned up the code for several drawing tools, as well as (gasp)
EraseTool.
- Changed the styling of the toolbar. I think it behaves nicer now, and
hopefully looks prettier too. (I need a designer, eh?)
- Also created a simple Makefile to easily compile our Hjson files to
JSON, which is what is actually used.

Working now:
- Configuration and theme loaded from external JSON.
- Which configuration and theme to load are specified via attributes on
the <note-taker> element.

Not working yet:
- Keyboard navigation/shortcuts.
- Attributes beyond just config-href and theme-href (like to allow a
user to use their own external toolbar, configured via HTML instead of/
in addition to JSON).
- Distinguishing an Apple Pencil or other stylus from regular touches/
mouse clicks.
- Multiple <note-taker> elements on one page *probably* doesn't work
right now, due to a limitation in PaperJS. I haven't tested this,
however. But hopefully it will be easy to get working.