Eliminate custom-builtin element NotetakerButton
This is a fairly minor refactor, just to eliminate NotetakerButton from
being a customized builtin element. Everything works exactly as in the
previous commit, but now all the buttons are just regular buttons, and
are configured via custom data-* attributes.

The next step now will be to accomplish two things: (1) allow for
multiple <note-taker> elements on the same page, and (2) split apart the
<note-taker> from the toolbar, so that the toolbar can be placed
separately, and also (3) do this in such a way that one toolbar can be
used for more than one <note-taker>. See the new notes in
ToolbarDesign.md for implementation details.